refactor: extract validation module from calculator (Phase 4a.8)
Move validateAttributeLimits, mastery cap validation (~250 lines),
and spell slot validation into calculatorModules/validation.ts.
This completes the 8-module extraction — the orchestrator is now
~300 lines. All snapshot tests pass with identical output.
